+++ DQM/Physics/src/TopDiLeptonOfflineDQM.h
@@ -113,7 +113,7 @@ namespace TopDiLeptonOffline {
     ///  6: passes conversion rejection and Isolation
     ///  7: passes the whole selection
     /// As described on https://twiki.cern.ch/twiki/bin/view/CMS/SimpleCutBasedEleID
-    double eidPattern_;
+    int eidPattern_;
     /// extra isolation criterion on electron
     StringCutObjectSelector<reco::GsfElectron>* elecIso_;
     /// extra selection on electrons
